requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
idna==2.9
idna-ssl==1.1.0
lxml==4.5.0
multidict==4.7.5
pkg-resources==0.0.0
python-dotenv==0.13.0
requests==2.23.0
soupsieve==2.0
typing-extensions==3.7.4.2
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
